backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML Libraries
openai-whisper>=20240930
requests==2.31.0
transformers==4.35.2
torch>=2.0.0

# Audio/Video Processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Note: You also need to install ffmpeg on your system
# - macOS: brew install ffmpeg
# - Ubuntu: sudo apt-get install ffmpeg
# - Windows: Download from https://ffmpeg.org/download.html